refactor(launcher): split history.ts into focused modules
- Extract types → history-types.ts, DB layer → history-db.ts, navigation → history-navigation.ts
- history.ts becomes a re-export barrel
- Fix findNextEpisode: deleted-file fallback now advances to next season
- Add isReadonlyWalRetryError (WAL-mode guard before read-write retry)
- Strengthen bun-sqlite.d.ts with generics on Statement/query/prepare
